requirements.txt
alpaca-py
pandas
sqlalchemy
matplotlib
